+++ lib/hcloud/collection.rb
@@ -66,7 +66,7 @@ module HCloud
     def inspect
       filters = filter_by.any? ? "filters: (#{filter_by.map { |k, v| "#{k} = #{v.inspect}" }.join(', ')})" : nil
       sort = sort_by ? "sort: (#{sort_by.join(', ')})" : nil
-      label_selector = label_selector ? "labels: (#{label_selector.map { |k, v| "#{k} = #{v.inspect}" }.join(', ')})" : nil
+      label_selector = label_selector ? "labels_selector: (#{label_selector})" : nil
 
       "#<#{self.class}#{[filters, sort, label_selector].compact.join(', ').presence&.prepend(' ')} page: #{page}, per_page: #{per_page}, total_entries: #{total_entries.inspect}>"
     end
@@ -80,7 +80,7 @@ module HCloud
         page: page,
         per_page: per_page,
         sort: sort_by,
-        label_selector: label_selector&.map { |k, v| "#{k}=#{v}" }&.join(","),
+        label_selector: label_selector,
         **filter_by,
       }
     end